requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rabbitmq==0.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
